Revert "refactored theme web views"
This reverts commit 2c542e8912b4795a10829f8cc37f5cf316554859.
